Sprint_4_version_2 - Удален класс Main - Отредактированы названия пакетов с POM и тестами. Немного изменена структура пакетов - URL вынесен как константа - В параметры добавлены недостающие вопросы и ответы.
