+++ src/test/java/ru/praktikum_services/qa_scooter/constants/Url.java
@@ -1,0 +1,5 @@
+package ru.praktikum_services.qa_scooter.constants;
+
+public class Url {
+    public static final String HOME_PAGE_URL = "https://qa-scooter.praktikum-services.ru/";
+}
