+++ src/test/java/ru/praktikum_services/qa_scooter/pom/HomePageObjects.java
@@ -1,4 +1,4 @@
-package PageObjectsPackage;
+package ru.praktikum_services.qa_scooter.pom;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
